fix(installer): BET-383 cycle-2 review — move stage table to shared, fix remount correctness
Block (fix-here): the cycle-1 line-count fix deleted stageLabel/stageIndex/
stageTotal from InstallerState to get production net negative, but that
made the status line WRONG after a mid-install remount — it always showed
"Checking the box · 1 of 6" regardless of the real stage, persisting until
the next stage transition (which may be the rest of the install), while
`stage` (sent to Copy diagnostics) held the correct id. Self-contradicting
state, and a regression against Scope D's "stop letting the renderer invent
stage labels."

Fix (the reviewer's preferred option): moved INSTALL_STAGES + STAGE_INDEX +
currentStageInfo out of src/main/installer/stageMapper.ts into a new
src/shared/installStages.ts. Both the main-process fold and the renderer
now derive the stage order/labels from the exact same source:

- stageMapper.ts re-exports INSTALL_STAGES/InstallStageId for its existing
  consumers (installer.ts, diagnostics.ts) — no other main-process file
  needed to change its imports. currentStageInfo is gone from stageMapper.ts
  entirely; its tests moved to src/shared/installStages.test.ts.
- handlers.ts's "stage" event and the installerState response go back to
  carrying just the raw stage id — the renderer now computes the label
  itself via the shared currentStageInfo, on both the live "stage" event
  and the remount-recovery path, so recovery is always correct instead of
  approximated.
- SshInstallStep.tsx drops its renderer-local INITIAL_STAGE_LABEL/
  TOTAL_STAGES constants in favour of `currentStageInfo(INITIAL_STAGE)`.

Net effect: production-code diff (excluding test files) is exactly 0
(324 add / 324 del) — within the "net 0 with a correct status line" bar
from the review — while fixing the correctness regression and removing
the last renderer-invented stage-label table.
