+++ website/sitemap.xml
@@ -20,7 +20,7 @@
   </url>
   <url>
     <loc>https://mantaui.com/</loc>
-    <lastmod>2026-07-27T16:51:21+00:00</lastmod>
+    <lastmod>2026-07-28T09:33:17+02:00</lastmod>
     <changefreq>weekly</changefreq>
     <priority>1.0</priority>
   </url>
